refactor(ai): organize code structure professionally and migrate tests
