+++ integration/icons.go
@@ -4,7 +4,7 @@ import "strings"
 
 var iconMap = map[string]string{
 	"git":            "󰊢",
-	"gh":             "󰊢",
+	"gh":             "",
 	"docker":         "",
 	"docker-compose": "",
 	"podman":         "󰡨",
@@ -21,8 +21,8 @@ var iconMap = map[string]string{
 	"npx":            "",
 	"pnpm":           "",
 	"pnpx":           "",
-	"bun":            "󰎣",
-	"bunx":           "󰎣",
+	"bun":            "",
+	"bunx":           "",
 	"yarn":           "",
 	"deno":           "",
 	"rust":           "",
@@ -104,14 +104,23 @@ var iconMap = map[string]string{
 	"just":           "󰷈",
 	"nmap":           "󰒍",
 	"ffmpeg":         "󰕼",
-	"ollama":         "󰚩",
+	"ollama":         "󰫢",
+	"ai":             "󰫢",
+	"llm":            "󰫢",
+	"openai":         "󰫢",
+	"groq":           "󰫢",
+	"gemini":         "󰫢",
+	"claude":         "󰫢",
+	"anthropic":      "󰫢",
+	"copilot":        "󰫢",
+	"chatgpt":        "󰚩",
 	"systemctl":      "󰒓",
 	"htop":           "󰍛",
 	"btop":           "󰍛",
 	"top":            "󰍛",
-	"tar":            "󰛪",
-	"zip":            "󰛪",
-	"unzip":          "󰛪",
+	"tar":            "",
+	"zip":            "",
+	"unzip":          "",
 	"alias":          "",
 	"history":        "",
 	"system":         "",
